packages/sinan-captcha/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sinan-captcha"
version = "0.1.40"
description = "Training and data engineering workspace for graphical captcha models."
readme = "README.md"
requires-python = ">=3.12"
# Keep the default install lightweight; command-specific heavy dependencies are loaded lazily.
dependencies = []

[project.optional-dependencies]
train = [
  # Non-backend-specific training/runtime dependencies shared by repo dev and setup-train.
  "numpy",
  "onnx",
  "onnxscript",
  "optuna",
  "pillow",
  "ultralytics",
]

[project.scripts]
sinan = "cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["cli", "group2_semantics", "project_metadata"]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = [
  "auto_train*",
  "autolabel*",
  "common*",
  "dataset*",
  "evaluate*",
  "exam*",
  "inference*",
  "materials*",
  "modeltest*",
  "ops*",
  "predict*",
  "release*",
  "solve*",
  "train*",
]

[tool.setuptools.package-data]
"auto_train" = [
  "resources/opencode/commands/*.md",
  "resources/opencode/skills/*/SKILL.md",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
files = ["src"]
strict = true
warn_unused_configs = true
